[추가]프로젝트 수정

+++ src/main/java/dnd/jackpot/stack/entity/StackDto.java
@@ -6,6 +6,6 @@ import lombok.NoArgsConstructor;
 
 @Data @AllArgsConstructor @NoArgsConstructor
 public class StackDto {
-	private int id=0;
+	private Long id=0L;
 	private EstackProgrammer stack;
 }
